Now using vectors, made better player movement

+++ src/main/java/org/zette/entities/Tree.java
@@ -1,7 +1,7 @@
 package org.zette.entities;
 
 public class Tree extends Entity {
-    public Tree(int x, int y) {
+    public Tree(double x, double y) {
         super(x, y, 50, 100, "/images/tree.png");
     }
 }
